ci(release_linux): change the runner to ubuntu-20.04
linuxdeployqt  (commit deebf70), build 38 built on 2022-12-15 18:06:02 UTC
ERROR: The host system is too new.
Please run on a system with a glibc version no newer than what comes with the oldest
currently still-supported mainstream distribution (Ubuntu Bionic), which is glibc 2.27.
This is so that the resulting bundle will work on most still-supported Linux distributions.
For more information, please see
https://github.com/probonopd/linuxdeployqt/issues/340